src/ft_transcendence_backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
redis==4.5.5
daphne==4.1.2
django-redis==5.2.0
Django==5.1.5
django-cors-headers==4.6.0
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
channels==4.1.0
channels-redis==4.1.0
opentelemetry-distro 
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-django==0.44b0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp-proto-grpc 
python-logging-loki==0.3.1
django-prometheus==2.3.1
pytest==8.3.5
requests==2.32.3
pytest-django==4.10.0
pytest-cov==6.0.0
web3==6.20.2
